from dev - prepare for 40.1

+++ python/mozbuild/mozbuild/base.py
@@ -25,7 +25,11 @@ from .mozconfig import (
     MozconfigLoadException,
     MozconfigLoader,
 )
+from .util import memoized_property
 from .virtualenv import VirtualenvManager
+
+
+_config_guess_output = []
 
 
 def ancestors(path):
@@ -40,8 +44,8 @@ def ancestors(path):
 def samepath(path1, path2):
     if hasattr(os.path, 'samefile'):
         return os.path.samefile(path1, path2)
-    return os.path.normpath(os.path.realpath(path1)) == \
-        os.path.normpath(os.path.realpath(path2))
+    return os.path.normcase(os.path.realpath(path1)) == \
+        os.path.normcase(os.path.realpath(path2))
 
 class BadEnvironmentException(Exception):
     """Base class for errors raised when the build environment is not sane."""
@@ -274,9 +278,6 @@ class MozbuildObject(ProcessExecutionMixin):
 
     @property
     def bindir(self):
-        import mozinfo
-        if mozinfo.os == "mac":
-            return os.path.join(self.topobjdir, 'dist', self.substs['MOZ_MACBUNDLE_NAME'], 'Contents', 'Resources')
         return os.path.join(self.topobjdir, 'dist', 'bin')
 
     @property
@@ -287,17 +288,47 @@ class MozbuildObject(ProcessExecutionMixin):
     def statedir(self):
         return os.path.join(self.topobjdir, '.mozbuild')
 
+    @memoized_property
+    def extra_environment_variables(self):
+        '''Some extra environment variables are stored in .mozconfig.mk.
+        This functions extracts and returns them.'''
+        from pymake.process import ClineSplitter
+        mozconfig_mk = os.path.join(self.topobjdir, '.mozconfig.mk')
+        env = {}
+        with open(mozconfig_mk) as fh:
+            for line in fh:
+                if line.startswith('export '):
+                    exports = ClineSplitter(line, self.topobjdir)[1:]
+                    for e in exports:
+                        if '=' in e:
+                            key, value = e.split('=')
+                            env[key] = value
+        return env
+
     def is_clobber_needed(self):
         if not os.path.exists(self.topobjdir):
             return False
         return Clobberer(self.topsrcdir, self.topobjdir).clobber_needed()
 
+    def have_winrm(self):
+        # `winrm -h` should print 'winrm version ...' and exit 1
+        try:
+            p = subprocess.Popen(['winrm.exe', '-h'],
+                                 stdout=subprocess.PIPE,
+                                 stderr=subprocess.STDOUT)
+            return p.wait() == 1 and p.stdout.read().startswith('winrm')
+        except:
+            return False
+
     def remove_objdir(self):
         """Remove the entire object directory."""
 
-        # We use mozfile because it is faster than shutil.rmtree().
-        # mozfile doesn't like unicode arguments (bug 818783).
-        rmtree(self.topobjdir.encode('utf-8'))
+        if sys.platform.startswith('win') and self.have_winrm():
+            subprocess.check_call(['winrm', '-rf', self.topobjdir])
+        else:
+            # We use mozfile because it is faster than shutil.rmtree().
+            # mozfile doesn't like unicode arguments (bug 818783).
+            rmtree(self.topobjdir.encode('utf-8'))
 
     def get_binary_path(self, what='app', validate_exists=True, where='default'):
         """Obtain the path to a compiled binary for this build configuration.
@@ -353,6 +384,11 @@ class MozbuildObject(ProcessExecutionMixin):
         if config_guess:
             return config_guess
 
+        # config.guess results should be constant for process lifetime. Cache
+        # it.
+        if _config_guess_output:
+            return _config_guess_output[0]
+
         p = os.path.join(topsrcdir, 'build', 'autoconf', 'config.guess')
 
         # This is a little kludgy. We need access to the normalize_command
@@ -362,7 +398,9 @@ class MozbuildObject(ProcessExecutionMixin):
         o = MozbuildObject(topsrcdir, None, None, None)
         args = o._normalize_command([p], True)
 
-        return subprocess.check_output(args, cwd=topsrcdir).strip()
+        _config_guess_output.append(
+                subprocess.check_output(args, cwd=topsrcdir).strip())
+        return _config_guess_output[0]
 
     def notify(self, msg):
         """Show a desktop notification with the supplied message
@@ -554,36 +592,45 @@ class MozbuildObject(ProcessExecutionMixin):
     def _make_path(self):
         baseconfig = os.path.join(self.topsrcdir, 'config', 'baseconfig.mk')
 
+        def is_xcode_lisense_error(output):
+            return self._is_osx() and 'Agreeing to the Xcode' in output
+
         def validate_make(make):
             if os.path.exists(baseconfig) and os.path.exists(make):
                 cmd = [make, '-f', baseconfig]
                 if self._is_windows():
                     cmd.append('HOST_OS_ARCH=WINNT')
                 try:
-                    subprocess.check_call(cmd, stdout=open(os.devnull, 'wb'),
-                        stderr=subprocess.STDOUT)
-                except subprocess.CalledProcessError:
-                    return False
-                return True
-            return False
+                    subprocess.check_output(cmd, stderr=subprocess.STDOUT)
+                except subprocess.CalledProcessError as e:
+                    return False, is_xcode_lisense_error(e.output)
+                return True, False
+            return False, False
 
+        xcode_lisense_error = False
         possible_makes = ['gmake', 'make', 'mozmake', 'gnumake']
 
         if 'MAKE' in os.environ:
             make = os.environ['MAKE']
-            if os.path.isabs(make):
-                if validate_make(make):
-                    return [make]
-            else:
-                possible_makes.insert(0, make)
+            possible_makes.insert(0, make)
 
         for test in possible_makes:
-            try:
-                make = which.which(test)
-            except which.WhichError:
-                continue
-            if validate_make(make):
+            if os.path.isabs(test):
+                make = test
+            else:
+                try:
+                    make = which.which(test)
+                except which.WhichError:
+                    continue
+            result, xcode_lisense_error_tmp = validate_make(make)
+            if result:
                 return [make]
+            if xcode_lisense_error_tmp:
+                xcode_lisense_error = True
+
+        if xcode_lisense_error:
+            raise Exception('Xcode requires accepting to the license agreement.\n'
+                'Please run Xcode and accept the license agreement.')
 
         if self._is_windows():
             raise Exception('Could not find a suitable make implementation.\n'
@@ -599,6 +646,9 @@ class MozbuildObject(ProcessExecutionMixin):
 
     def _is_windows(self):
         return os.name in ('nt', 'ce')
+
+    def _is_osx(self):
+        return 'darwin' in str(sys.platform).lower()
 
     def _spawn(self, cls):
         """Create a new MozbuildObject-derived class instance from ourselves.
